nitial project setup with base dependencies + Getting to understand how to setup Cats API
1º - Create the application, and figure out what dependencies I'll need to do the challenge.
Although I understand Kotlin and took a few classes on the programming language during my undergraduate studies, I've never implemented any API in it, since all my mobile development projects were designed to be coded in Flutter. So it was necessary to do some global research to understand what I would need. Then, after researching, I realized that I would need some plugins like ksp and dependencies like retrofit, moshi, room.

So, after the research, I started preparing all the code by creating a breed class with all the necessary variables, an interface to get the races and also an object that will make the call to the API.

At the end, I created a view so that I could see if the call was made correctly. Obviously, the main point is to have a way of adding a race to the favorites later on

Note: The APY key is public at the moment, but it's something that will be fixed later. The main focus was to learn how to work with API's in Kotlin, since I've never worked with it before